time and cost changed to int, all actions are now checked for feasibility when new action is added, currentGroups changed to set in group generator, all subsets are now checked in group generation, bug with station storage fixed in VehicleGroupAssignmentSolver
Former-commit-id: 2410d0f7e395599b0e2232f6d1754882c6c75a2f

+++ src/main/java/cz/cvut/fel/aic/amodsim/ridesharing/vga/calculations/OptimalVehiclePlanFinder.java
@@ -23,6 +23,6 @@ public abstract class OptimalVehiclePlanFinder<V extends IOptimalPlanVehicle> {
 	}
 	
 	public abstract Plan<V> getOptimalVehiclePlanForGroup(V vehicle, List<VGAVehiclePlanAction> actions, 
-			double startTime, boolean ignoreTime);
+			int startTime, boolean ignoreTime);
 	
 }
